+++ cmd/app.go
@@ -1,42 +1,43 @@
 package main
 
 import (
-  "fmt"
+	"fmt"
 
-  _ "github.com/kalilventura/vehicle-management-payment/cmd/docs" // generated docs
-  "github.com/kalilventura/vehicle-management-payment/internal/shared/domain/entities"
-  "github.com/labstack/echo/v4"
-  logger "github.com/sirupsen/logrus"
-  "github.com/swaggo/echo-swagger" // echo-swagger middleware
+	_ "github.com/kalilventura/vehicle-management-payment/cmd/docs" // generated docs
+	"github.com/kalilventura/vehicle-management-payment/internal/shared/domain/entities"
+	"github.com/labstack/echo/v4"
+	logger "github.com/sirupsen/logrus"
+	echoSwagger "github.com/swaggo/echo-swagger" // echo-swagger middleware
 )
 
 func handleRoutes(application *echo.Echo, modules []entities.HTTPModule) {
-  logger.Info("🚀 Initializing HTTP router and registering routes...")
-  groups := map[string]*echo.Group{}
+	logger.Info("🚀 Initializing HTTP router and registering routes...")
+	groups := map[string]*echo.Group{}
 
-  for _, module := range modules {
-    for _, controller := range module.GetControllers() {
-      bind := controller.GetBind()
+	for _, module := range modules {
+		for _, controller := range module.GetControllers() {
+			bind := controller.GetBind()
 
-      group, found := groups[bind.Version]
-      if !found {
-        group = application.Group(bind.Version)
-        groups[bind.Version] = group
-      }
-      message := fmt.Sprintf("router registered: [%s] %s", bind.Method, bind.GetFullPath())
-      logger.Info(message)
-      group.Add(bind.Method, bind.RelativePath, controller.Execute)
-    }
-  }
+			group, found := groups[bind.Version]
+			if !found {
+				group = application.Group(bind.Version)
+				groups[bind.Version] = group
+			}
+			message := fmt.Sprintf("router registered: [%s] %s", bind.Method, bind.GetFullPath())
+			logger.Info(message)
+			group.Add(bind.Method, bind.RelativePath, controller.Execute)
+		}
+	}
 }
 
 func handleSwagger(application *echo.Echo) {
-  logger.Info("registering swagger")
-  application.GET("/swagger/*", echoSwagger.WrapHandler)
+	logger.Info("registering swagger")
+	application.GET("/swagger/*", echoSwagger.WrapHandler)
 }
 
 func SetupServer(modules []entities.HTTPModule) *echo.Echo {
-  application := echo.New()
-  handleRoutes(application, modules)
-  return application
+	application := echo.New()
+	handleRoutes(application, modules)
+	handleSwagger(application)
+	return application
 }