+++ internal/payments/domain/entities/update_payment.go
@@ -1,0 +1,7 @@
+package entities
+
+type UpdatePayment struct {
+	GatewayPaymentID   string
+	Status             string
+	PaymentTransaction *PaymentTransaction
+}