+++ internal/payments/infrastructure/repositories/mappers/payment_mapper.go
@@ -6,10 +6,17 @@ import (
 )
 
 func MapToEntity(payment *entities.Payment) *models.GormPayment {
+	transaction := models.GormPaymentTransaction{
+		GatewayTransactionID: payment.PaymentTransaction.GatewayTransactionID,
+		Status:               payment.PaymentTransaction.Status,
+		ResponseCode:         payment.PaymentTransaction.ResponseCode,
+		ResponseMessage:      payment.PaymentTransaction.ResponseMessage,
+		RawResponse:          payment.PaymentTransaction.RawResponse,
+	}
 	return &models.GormPayment{
-		VehicleID: payment.VehicleID,
-		Cpf:       payment.Cpf,
-		Amount:    payment.Amount,
-		CreatedAt: payment.CreatedAt,
+		VehicleID:    payment.VehicleID,
+		Cpf:          payment.Cpf,
+		Amount:       payment.Amount,
+		Transactions: []models.GormPaymentTransaction{transaction},
 	}
 }