+++ internal/shared/domain/entities/settings.go
@@ -6,10 +6,6 @@ type Settings struct {
 	Port int
 }
 
-func NewSettings(port int) *Settings {
-	return &Settings{port}
-}
-
 func (s Settings) GetPort() string {
 	return fmt.Sprintf(":%d", s.Port)
 }